requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
Django==3.0.5
django-filter==2.2.0
django-pdb==0.6.2
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.5.0
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.1.2
pipenv==2018.11.26
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
zipp==3.1.0
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
Django==3.0.5
django-filter==2.2.0
django-pdb==0.6.2
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.5.0
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.1.2
pipenv==2018.11.26
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
zipp==3.1.0
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
Django==3.0.5
django-filter==2.2.0
django-pdb==0.6.2
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.5.0
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.1.2
pipenv==2018.11.26
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
zipp==3.1.0
